requirements.txt
# Ottawa City Services RAG Chatbot - Dependencies
# Core RAG Pipeline
sentence-transformers==2.2.2
chromadb==0.4.15
langchain==0.1.0
langchain-community==0.0.10

# LLM and API
groq==0.4.1
openai==1.3.0

# Web Scraping
scrapy==2.11.0
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.0

# Data Processing
pandas==2.1.3
numpy==1.24.3
python-dotenv==1.0.0

# Web Interface
gradio==4.8.0
streamlit==1.28.0

# Utilities
tqdm==4.66.1
python-multipart==0.0.6

# Compatible Dependencies (Fixed Versions)
protobuf==3.20.3
pydantic==2.5.0
typing-extensions==4.8.0

# Development and Testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
pre-commit==3.5.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.0

# Deployment
uvicorn==0.24.0
fastapi==0.104.1
gunicorn==21.2.0

# Optional: Advanced Features
redis==5.0.1
celery==5.3.4